Moved util methods in DicomStreams to companion object DicomStreamOps. Deleted DicomStreams.
